Move msa datatypes into separate file
